Font fix
Adds an if statement that checks if Open Sans is available before making figures and switches to a properly-sized Arial text if the computer is sans Open Sans. Formerly, the computer would switch to the font whose name was most similar to Open Sans and this would not properly format within the figure.